requirements.txt
requests>=2.26.0
dash>=2.3.1
matplotlib>=3.5.2
numpy>=1.23.1
pandas==1.4.3
scipy>=1.9.0
seaborn>=0.11.2
pytest>=7.1.2
openpyxl>=3.0.10
dash_bootstrap_components>=1.2.1
pyyaml>=6.0
typer>=0.7.0
zhutils @ git+https://github.com/mikewellmeansme/ZhUtils